pyproject.toml
[project]
name = "scratch-inspector"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "altair<5",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "plotly>=6.6.0",
    "scipy>=1.17.1",
    "streamlit>=1.19.0",
]
